+++ Magic/src/main/java/com/elmakers/mine/bukkit/action/builtin/ParallelAction.java
@@ -11,7 +11,21 @@ import com.elmakers.mine.bukkit.api.spell.SpellResult;
 
 public class ParallelAction extends CompoundAction
 {
-    private List<ActionContext> remaining = null;
+    private static class SubAction {
+        private ActionContext action;
+        private CastContext context;
+
+        public SubAction(ActionContext action, CastContext context) {
+            this.action = action;
+            this.context = new com.elmakers.mine.bukkit.action.CastContext(context);
+        }
+
+        public SpellResult perform() {
+            return action.perform(context);
+        }
+    }
+
+    private List<SubAction> remaining = null;
 
     @Override
     public void reset(CastContext context) {
@@ -19,9 +33,10 @@ public class ParallelAction extends CompoundAction
         ActionHandler actions = getHandler("actions");
         if (actions != null)
         {
-            remaining = new ArrayList<>(actions.getActions());
-            for (ActionContext action : remaining) {
+            remaining = new ArrayList<>();
+            for (ActionContext action : actions.getActions()) {
                 action.getAction().reset(context);
+                remaining.add(new SubAction(action, context));
             }
         }
     }
@@ -40,13 +55,13 @@ public class ParallelAction extends CompoundAction
     public SpellResult perform(CastContext context) {
         SpellResult result = SpellResult.NO_ACTION;
         int startingWork = context.getWorkAllowed();
-        List<ActionContext> subActions = new ArrayList<>(remaining);
+        List<SubAction> subActions = new ArrayList<>(remaining);
         remaining.clear();
         context.setWorkAllowed(0);
         int splitWork = Math.max(1, startingWork / subActions.size());
-        for (ActionContext action : subActions) {
+        for (SubAction action : subActions) {
             context.setWorkAllowed(context.getWorkAllowed() + splitWork);
-            SpellResult actionResult = action.perform(context);
+            SpellResult actionResult = action.perform();
             context.addResult(actionResult);
             if (actionResult.isStop()) {
                 remaining.add(action);